# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Guardian Server API
  version: 1.0.0
  description: |
    Guardian 后端 REST API 文档。
    SQLite 持久化 admins/teachers，rooms/clients 全量在内存。
    WebSocket 协议见 `docs/asyncapi.yaml`。
...
servers:
  - url: http://{host}:3847
    variables:
      host:
        default: localhost
...
paths:
...
  /api/teacher/login:
    post:
      tags: [Teacher]
      summary: 教师登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherLoginResponse'
        '401':
          description: 工号或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/teacher/password:
    put:
      tags: [Teacher]
      summary: 修改当前教师密码
      security:
        - XTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherPasswordRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkOnly'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    XTokenAuth:
      type: apiKey
      in: header
      name: x-token
...
  responses:
...
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OkOnly:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [ok, msg]
      properties:
        ok:
          type: boolean
          enum: [false]
        msg:
          type: string
...
    TeacherLoginRequest:
      type: object
      required: [staffId, password]
      properties:
        staffId:
          type: string
        password:
          type: string
    TeacherLoginResponse:
      type: object
      required: [ok, token, teacherId, staffId, name]
      properties:
        ok:
          type: boolean
          enum: [true]
        token:
          type: string
        teacherId:
          type: string
        staffId:
          type: string
        name:
          type: string
    TeacherPasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
...